Rationalize RangeSet iteration order to eliminate need to externally sort large lists of Range objects.
This change allow client code to provide contigs up-front upon RangeSet construction, allowing us to have a built-in natural sort order over all ranges, based on the pos_in_fasta of the contig.  If the contigs argument is not provided, the sort order falls back to the alphabetical order of contig names, as before.

IF contigs is provided, an exception will be raised if a contig name in `ranges` cannot be matched with a record in `contigs`.

A test is added for the sortedness of RangeSet; vestigial code for sorting ranges is removed.

`partition` is now also guaranteed to yield ranges in a sorted order (since the underlying RangeSet has no overlapping intervals and itself yields intervals in sorted order now); test and docstring updated accordingly.

PiperOrigin-RevId: 193716863